// js_modules/dagit/src/schedules/types/ScheduleRootQuery.ts
export interface ScheduleRootQuery_scheduleOrError_RunningSchedule_attemptList_run {
  __typename: "PipelineRun";
  runId: string;
  status: PipelineRunStatus;
}

export interface ScheduleRootQuery_scheduleOrError_RunningSchedule_attemptList {
  __typename: "ScheduleAttempt";
  time: number;
  jsonResult: string;
  status: ScheduleAttemptStatus;
  run: ScheduleRootQuery_scheduleOrError_RunningSchedule_attemptList_run | null;
}

export interface ScheduleRootQuery_scheduleOrError_RunningSchedule {
  __typename: "RunningSchedule";
  scheduleDefinition: ScheduleRootQuery_scheduleOrError_RunningSchedule_scheduleDefinition;
  logsPath: string;
  ticks: ScheduleRootQuery_scheduleOrError_RunningSchedule_ticks[];
  runsCount: number;
  runs: ScheduleRootQuery_scheduleOrError_RunningSchedule_runs[];
  stats: ScheduleRootQuery_scheduleOrError_RunningSchedule_stats;
  ticksCount: number;
  status: ScheduleStatus;
  ticksList: ScheduleRootQuery_scheduleOrError_RunningSchedule_ticksList[];
  attemptList: ScheduleRootQuery_scheduleOrError_RunningSchedule_attemptList[];
}

export interface ScheduleRootQuery_scheduleOrError_ScheduleNotFoundError {
  __typename: "ScheduleNotFoundError";
  message: string;
}

export interface ScheduleRootQuery_scheduleOrError_PythonError {
  __typename: "PythonError";
  message: string;
  stack: string[];
}

export type ScheduleRootQuery_scheduleOrError = ScheduleRootQuery_scheduleOrError_RunningSchedule | ScheduleRootQuery_scheduleOrError_ScheduleNotFoundError | ScheduleRootQuery_scheduleOrError_PythonError;

export interface ScheduleRootQuery {
  scheduleOrError: ScheduleRootQuery_scheduleOrError;
}